IKHSS7-38210: msm: ipc: Add support to load and unload the modem
When all the user-space clients are killed, modem gets unloaded and goes
to silent reset. When the user-space clients come up again, modem resets
its side of the RPCRPY_CNTL SMD channel. This leaves the SMD channel in
stale state. Hence modem starts reading stale data from the SMD channel
and crashes. So as long as the IPC Router uses the SMD channel, cast a
vote to PIL to ensure that the modem is loaded.

Change-Id: I27a49b8ba3f953df316c6a2f171e0d0f0c7a9355
CRs-Fixed: 359773
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@codeaurora.org>
Signed-off-by: Ken Hu <jian-wu.hu@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/443189
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Christopher Fries <C.Fries@motorola.com>
Reviewed-by: Nir Darmoni <Darmoni@motorola.com>
Reviewed-by: Igor Kovalenko <Igor.Kovalenko@motorola.com>