IKHSS7-43056: fix iommu page fault when play video
iommu: Unexpected IOMMU page fault! happens at the end of video playback,
because mdp4_free_writeback_buf is freed when it is still in use. This
issue was exposed when enabling BLT mode for 1080p/720p portrait mode
video plabback.

This is to integrate a patch from qcom in case 899668/CR # 375073 to fix
it.

Change-Id: Iae12b1be33a8f76b2ec92627acedc32cfc624427
Signed-off-by: Limei Wang <e12499@motorola.com>